+++ lib/store/settings.dart
@@ -80,6 +80,9 @@ class SettingsRepository extends DisposableInterface
   Future<void> setPopupsEnabled(bool enabled) =>
       _settingsLocal.setPopupsEnabled(enabled);
 
+  @override
+  Future<void> setLocale(String locale) => _settingsLocal.setLocale(locale);
+
   /// Initializes [MediaSettingsHiveProvider.boxEvents] subscription.
   Future<void> _initMediaSubscription() async {
     _mediaSubscription = StreamIterator(_mediaLocal.boxEvents);