+++ lib/ui/page/call/component/desktop.dart
@@ -37,6 +37,7 @@ import '../widget/scaler.dart';
 import '../widget/tooltip_button.dart';
 import '../widget/video_view.dart';
 import '/domain/model/ongoing_call.dart';
+import '/l10n/l10n.dart';
 import '/routes.dart';
 import '/themes.dart';
 import '/ui/page/home/page/chat/widget/chat_item.dart';
@@ -272,7 +273,7 @@ Widget desktopCall(
                   width: 240,
                   height: 110,
                   child: HintWidget(
-                    text: 'label_hint_drag_n_drop_video'.tr,
+                    text: 'label_hint_drag_n_drop_video'.l10n,
                     onTap: c.isHintDismissed.toggle,
                   ),
                 ),
@@ -809,7 +810,7 @@ Widget desktopCall(
               padding: const EdgeInsets.only(top: 8, right: 8),
               child: TooltipButton(
                 verticalOffset: 8,
-                hint: 'btn_call_settings'.tr,
+                hint: 'btn_call_settings'.l10n,
                 onTap: () => c.openSettings(context),
                 child: SvgLoader.asset(
                   'assets/icons/settings.svg',
@@ -1141,12 +1142,12 @@ Widget _titleBar(
         String state = c.state.value == OngoingCallState.active
             ? c.duration.value.localizedString()
             : c.state.value == OngoingCallState.joining
-                ? 'label_call_joining'.tr
+                ? 'label_call_joining'.l10n
                 : isOutgoing
-                    ? 'label_call_calling'.tr
+                    ? 'label_call_calling'.l10n
                     : c.withVideo == true
-                        ? 'label_video_call'.tr
-                        : 'label_audio_call'.tr;
+                        ? 'label_video_call'.l10n
+                        : 'label_audio_call'.l10n;
 
         return Container(
           key: const ValueKey('TitleBar'),
@@ -1187,7 +1188,7 @@ Widget _titleBar(
                             child: ConstrainedBox(
                               constraints: const BoxConstraints(maxWidth: 160),
                               child: Text(
-                                c.chat.value?.title.value ?? ('.'.tr * 3),
+                                c.chat.value?.title.value ?? ('dot'.l10n * 3),
                                 style: context.textTheme.bodyText1?.copyWith(
                                   fontSize: 17,
                                   color: const Color(0xFFBBBBBB),
@@ -1249,8 +1250,8 @@ Widget _titleBar(
                           TooltipButton(
                             onTap: c.toggleRemoteVideos,
                             hint: c.isRemoteVideoEnabled.value
-                                ? 'btn_call_disable_video'.tr
-                                : 'btn_call_enable_video'.tr,
+                                ? 'btn_call_disable_video'.l10n
+                                : 'btn_call_enable_video'.l10n,
                             child: Icon(
                               c.isRemoteVideoEnabled.value
                                   ? Icons.videocam
@@ -1262,7 +1263,7 @@ Widget _titleBar(
                           const SizedBox(width: 12),
                           TooltipButton(
                             onTap: () => c.openAddMember(context),
-                            hint: 'btn_add_participant'.tr,
+                            hint: 'btn_add_participant'.l10n,
                             child: Padding(
                               padding: const EdgeInsets.only(top: 2),
                               child: SvgLoader.asset(
@@ -1276,7 +1277,7 @@ Widget _titleBar(
                         ],
                         TooltipButton(
                           onTap: () => c.openSettings(context),
-                          hint: 'btn_call_settings'.tr,
+                          hint: 'btn_call_settings'.l10n,
                           child: SvgLoader.asset(
                             'assets/icons/settings.svg',
                             color: const Color(0xFFBBBBBB),
@@ -1287,8 +1288,8 @@ Widget _titleBar(
                         TooltipButton(
                           onTap: c.toggleFullscreen,
                           hint: c.fullscreen.value
-                              ? 'btn_fullscreen_exit'.tr
-                              : 'btn_fullscreen_enter'.tr,
+                              ? 'btn_fullscreen_exit'.l10n
+                              : 'btn_fullscreen_enter'.l10n,
                           child: SvgLoader.asset(
                             'assets/icons/fullscreen_${c.fullscreen.value ? 'exit' : 'enter'}.svg',
                             width: 14,
@@ -1446,8 +1447,8 @@ Widget _primaryVideo(
                         if (participant.source == MediaSourceKind.Device)
                           ContextMenuButton(
                             label: fit == null || fit == BoxFit.cover
-                                ? 'btn_call_do_not_cut_video'.tr
-                                : 'btn_call_cut_video'.tr,
+                                ? 'btn_call_do_not_cut_video'.l10n
+                                : 'btn_call_cut_video'.l10n,
                             onPressed: () {
                               c.rendererBoxFit[
                                       participant.video.value!.track.id()] =
@@ -1464,14 +1465,14 @@ Widget _primaryVideo(
                           ),
                         if (mayDragVideo)
                           ContextMenuButton(
-                            label: 'btn_call_center_video'.tr,
+                            label: 'btn_call_center_video'.l10n,
                             onPressed: () => c.center(participant),
                           ),
                       ],
                       ContextMenuButton(
                         label: participant.video.value?.isEnabled == true
-                            ? 'btn_call_disable_video'.tr
-                            : 'btn_call_enable_video'.tr,
+                            ? 'btn_call_disable_video'.l10n
+                            : 'btn_call_enable_video'.l10n,
                         onPressed: () =>
                             c.toggleRendererEnabled(participant.video),
                       ),
@@ -1574,13 +1575,13 @@ Widget _secondaryVideo(
                           participant.source != MediaSourceKind.Display) &&
                       participant.video.value?.isEnabled == true)
                     ContextMenuButton(
-                      label: 'btn_call_center_video'.tr,
+                      label: 'btn_call_center_video'.l10n,
                       onPressed: () => c.center(participant),
                     ),
                   ContextMenuButton(
                     label: participant.video.value?.isEnabled == true
-                        ? 'btn_call_disable_video'.tr
-                        : 'btn_call_enable_video'.tr,
+                        ? 'btn_call_disable_video'.l10n
+                        : 'btn_call_enable_video'.l10n,
                     onPressed: () => c.toggleRendererEnabled(participant.video),
                   )
                 ],
@@ -1769,7 +1770,7 @@ Widget _secondaryTarget(
                                 ),
                           const SizedBox(height: 5),
                           Text(
-                            'btn_call_drop_video_here'.tr,
+                            'btn_call_drop_video_here'.l10n,
                             style: context.textTheme.subtitle1?.copyWith(
                               color: const Color(0xFFBBBBBB),
                             ),