+++ lib/ui/page/home/tab/chats/view.dart
@@ -24,6 +24,7 @@ import '/domain/model/chat_call.dart';
 import '/domain/model/chat_item.dart';
 import '/domain/repository/chat.dart';
 import '/domain/repository/user.dart';
+import '/l10n/l10n.dart';
 import '/routes.dart';
 import '/ui/page/call/widget/animated_dots.dart';
 import '/ui/page/home/page/chat/controller.dart' show ChatCallFinishReasonL10n;
@@ -46,7 +47,7 @@ class ChatsTabView extends StatelessWidget {
       builder: (ChatsTabController c) {
         return Scaffold(
           appBar: AppBar(
-            title: Text('label_chats'.tr),
+            title: Text('label_chats'.l10n),
             bottom: PreferredSize(
               preferredSize: const Size.fromHeight(0.5),
               child: Container(
@@ -67,7 +68,7 @@ class ChatsTabView extends StatelessWidget {
           body: Obx(
             () => c.chatsReady.value
                 ? c.chats.isEmpty
-                    ? Center(child: Text('label_no_chats'.tr))
+                    ? Center(child: Text('label_no_chats'.l10n))
                     : ContextMenuInterceptor(
                         child: ListView(
                           controller: ScrollController(),
@@ -117,8 +118,8 @@ class ChatsTabView extends StatelessWidget {
                     ),
                     const SizedBox(width: 2),
                     Text(typings.length > 1
-                        ? 'label_typings'.tr
-                        : 'label_typing'.tr),
+                        ? 'label_typings'.l10n
+                        : 'label_typing'.l10n),
                     const AnimatedDots(color: Colors.black)
                   ],
                 ),
@@ -127,14 +128,14 @@ class ChatsTabView extends StatelessWidget {
           } else if (chat.lastItem != null) {
             if (chat.lastItem is ChatCall) {
               var item = chat.lastItem as ChatCall;
-              String description = 'label_chat_call_ended'.tr;
+              String description = 'label_chat_call_ended'.l10n;
               if (item.finishedAt == null && item.finishReason == null) {
                 subtitle = [
                   Padding(
                     padding: const EdgeInsets.fromLTRB(0, 4, 0, 4),
                     child: ElevatedButton(
                       onPressed: () => c.joinCall(chat.id),
-                      child: Text('btn_chat_join_call'.tr),
+                      child: Text('btn_chat_join_call'.l10n),
                     ),
                   ),
                 ];
@@ -156,18 +157,18 @@ class ChatsTabView extends StatelessWidget {
               var desc = StringBuffer();
 
               if (!chat.isGroup && item.authorId == c.me) {
-                desc.write('${'label_you'.tr}: ');
+                desc.write('${'label_you'.l10n}: ');
               }
 
               if (item.text != null) {
                 desc.write(item.text!.val);
                 if (item.attachments.isNotEmpty) {
                   desc.write(
-                      ' [${item.attachments.length} ${'label_attachments'.tr}]');
+                      ' [${item.attachments.length} ${'label_attachments'.l10n}]');
                 }
               } else if (item.attachments.isNotEmpty) {
                 desc.write(
-                    '[${item.attachments.length} ${'label_attachments'.tr}]');
+                    '[${item.attachments.length} ${'label_attachments'.l10n}]');
               }
 
               subtitle = [
@@ -203,7 +204,7 @@ class ChatsTabView extends StatelessWidget {
                 child: ElevatedButton(
                   onPressed: () => c.joinCall(chat.id),
                   child: Text(
-                    'btn_chat_join_call'.tr,
+                    'btn_chat_join_call'.l10n,
                     maxLines: 2,
                     overflow: TextOverflow.ellipsis,
                   ),
@@ -220,13 +221,13 @@ class ChatsTabView extends StatelessWidget {
             actions: [
               ContextMenuButton(
                 key: const Key('ButtonHideChat'),
-                label: 'btn_hide_chat'.tr,
+                label: 'btn_hide_chat'.l10n,
                 onPressed: () => c.hideChat(chat.id),
               ),
               if (chat.isGroup)
                 ContextMenuButton(
                   key: const Key('ButtonLeaveChat'),
-                  label: 'btn_leave_chat'.tr,
+                  label: 'btn_leave_chat'.l10n,
                   onPressed: () => c.leaveChat(chat.id),
                 ),
             ],