+++ lib/ui/page/home/widget/user_search_bar/view.dart
@@ -20,6 +20,7 @@ import 'package:material_floating_search_bar/material_floating_search_bar.dart';
 
 import '/domain/model/user.dart';
 import '/domain/repository/user.dart';
+import '/l10n/l10n.dart';
 import '/ui/page/home/widget/avatar.dart';
 import 'controller.dart';
 
@@ -59,7 +60,7 @@ class UserSearchBar extends StatelessWidget {
       init: UserSearchBarController(Get.find()),
       builder: (UserSearchBarController c) => Obx(
         () => FloatingSearchBar(
-          hint: 'label_search'.tr,
+          hint: 'label_search'.l10n,
           controller: searchController,
           scrollPadding: const EdgeInsets.only(top: 16, bottom: 56),
           transitionDuration: const Duration(milliseconds: 500),
@@ -93,7 +94,7 @@ class UserSearchBar extends StatelessWidget {
                                 ? [
                                     ListTile(
                                         title:
-                                            Text('label_search_not_found'.tr))
+                                            Text('label_search_not_found'.l10n))
                                   ]
                                 : c.searchResults
                                     .map((e) => _user(e, c))
@@ -105,12 +106,13 @@ class UserSearchBar extends StatelessWidget {
                                 SizedBox(
                                   height: 100,
                                   child: Center(
-                                    child: Text('label_search_hint'.tr),
+                                    child: Text('label_search_hint'.l10n),
                                   ),
                                 )
                               ]
                             : [
-                                ListTile(title: Text('label_search_recent'.tr)),
+                                ListTile(
+                                    title: Text('label_search_recent'.l10n)),
                                 ...c.recentSearchResults
                                     .map((e) => _user(e, c))
                                     .toList()