+++ lib/util/message_popup.dart
@@ -15,8 +15,8 @@
 // <https://www.gnu.org/licenses/agpl-3.0.html>.
 
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 
+import '/l10n/l10n.dart';
 import '/routes.dart';
 import 'localized_exception.dart';
 
@@ -28,12 +28,12 @@ class MessagePopup {
     await showDialog(
       context: router.context!,
       builder: (context) => AlertDialog(
-        title: Text('label_error'.tr),
+        title: Text('label_error'.l10n),
         content: Text(message),
         actions: [
           TextButton(
             onPressed: () => Navigator.of(router.context!).pop(),
-            child: Text('btn_ok'.tr),
+            child: Text('btn_ok'.l10n),
           )
         ],
       ),
@@ -52,12 +52,12 @@ class MessagePopup {
           actions: [
             TextButton(
               key: const Key('AlertNoButton'),
-              child: Text('label_are_you_sure_no'.tr),
+              child: Text('label_are_you_sure_no'.l10n),
               onPressed: () => Navigator.pop(context, false),
             ),
             TextButton(
               key: const Key('AlertYesButton'),
-              child: Text('label_are_you_sure_yes'.tr),
+              child: Text('label_are_you_sure_yes'.l10n),
               onPressed: () => Navigator.pop(context, true),
             ),
           ],